refactored admin to be more structured
