+++ aldryn_forms/admin/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+from . import formdata
+from . import formsubmission
